refactor(door) refactor door class as separate class (#11)
