+++ Program.cs
@@ -5,18 +5,5 @@
         Console.WriteLine("Welcome to the Car Door Counter!");
     }
 
-    public class Car
-    {
-        private readonly int _numberOfDoors = 2;
-
-        public int NumberOfDoors
-        {
-            get { return _numberOfDoors; }
-        }
-
-        public bool GetNumberOfDoors()
-        {
-            return NumberOfDoors > 1 && NumberOfDoors < 5;
-        }
-    }
+    
 }
